Combined NoExisting methods and renamed them.
Processed review comment: the checkNoExistingExcludedFields and checkNoExistingExcludedFields now have their shared logic extracted to a third method. The naming has also been adjusted: check has been replaced with assert to make more expressive that this is an expected state of the application that will result in a thrown exception if it is not true. Also applied this change to checkNoNonnullFields to be consistent.